git-compat-util(msvc): C11 does not imply support for zero-sized arrays
This seems to be required to define `FLEX_ARRAY` in a manner that MSVC
in C11 mode accepts. Without this fix, building Git for Windows'
experimental FSCache code would fail thusly:

	error C2229: struct 'heap_fsentry' has an illegal zero-sized array

The reason is that we need to extend a struct with a flexible array
member that holds a file name, to the maximally allowed length of a file
name, so that we can allocate one instance of that struct on the heap
and reuse it for several file names.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>