git-svn: fix searching for parent branch when rewriteRoot is used
When setting `rewriteRoot` is used, the value of `git-svn-id` (in commit
message) is prefixed with the value of `rewriteRoot`. When searching
for parent of a new branch, if URL of a svn remote is not found in
`git-svn-id` (see sub `resolve_local_globs`), then git-svn will stop.

This fix replaced this URL with the value of `rewriteRoot`, so that git-svn
could find parent branch when setting `rewriteRoot` is used.

Signed-off-by: Dung Saga <dung.saga@gmail.com>