Provide config option to expect files outside sparse patterns
Typically with sparse checkouts, we expect files outside the sparsity
patterns to be marked as SKIP_WORKTREE and be missing from the working
tree.  There are a few edge cases where this can be violated and cause
confusion, so in a sparse checkout, Git will check whether SKIP_WORKTREE
files are present and clear the SKIP_WORKTREE bit if they are (see
commit 11d46a399d ("repo_read_index: clear SKIP_WORKTREE bit from files
present in worktree", 2022-01-06).

There is currently an effort code-named `vfsd` (see
https://lore.kernel.org/git/20220207190320.2960362-1-jonathantanmy@google.com/
for details) to implement a Git-aware virtual file system (VFS) layer
(similar in spirit to the VFS for Git project) that can be used to turn
this expectation on its head: all files are considered present in the
working copy, though they are not vivified until actually accessed.

With such a VFS, most of the files do not match the sparsity patterns at
first, and the VFS layer automatically updates the sparsity patterns to
add more files whenever files are written.

Add an option, sparse.expectFilesOutsideOfPatterns, which can be used by
this VFS to signal to Git to not check for the presence of SKIP_WORKTREE
files.

Signed-off-by: Elijah Newren <newren@gmail.com>