Refactor: list_cmds()
When we are looking at "if/else if" statements that uses a return value
of a function call with some arguments that repeats and only a few
changes, is more "clean" to set variables that express the return value
of each function call. This way instead of spend time looking at the
function body to understand what the function does, we look the name of
the variable that holds the return value of the function.

Signed-off-by: thiago16fca <thiago16fca@outlook.com>