t3200: fix antipatterns in existing branch tests
Fix issues in t3200 branch tests that, if copied, might catch new
contributors out:

Use test_config to show that config state is not being intentionally
left to spill over into other tests.

Use test_cmp_config instead of git config in subshells, so that
git's error code is not lost if/when an unexpected error occurs.

Use output redirection and later content checking instead of
subshells, so that git's error code is not lost if/when an
unexpected error occurs.

Eliminate local-fetch-avoiding optimization as it is error-prone (it
is easy to check the wrong thing), hides segfaults, and yields only
a marginal performance improvement given the fetch is local.

Introduce local helper test_set_remote to simplify the common local
pattern of setting up a remote via config.

These changes do not aspire to ensure that no config or state spills
over between tests, but rather to use the helpers that avoid this by
default, to be clearer that any state spillover is incidental.

In cases where state previously "leaked" and no longer does due to a
more state-spillover-concious helper, some checking for downstream
dependencies was done (none found), but correctness was naively
assumed on the basis of tests still passing.

Signed-off-by: Tao Klerks <tao@klerks.biz>