Fix `__GNUC_MINOR__` spelling mistakes
[Please refer to **The common predefined macros are GNU C extensions**]
(https://gcc.gnu.org/onlinedocs/cpp/Common-Predefined-Macros.html)

Recently, the work needs to be developed on the SUSE 9.2 system (the
system released in 2004), the gcc version is 3.3.4, and the
"undefined reference to __builtin_ctzll" error occurred when compiling
git-2.1.0. The investigation found that the __GUNC_MINOR__ macro was
misspelled.

Presumably everyone will not use such an old system,
so this error has not been found.

Signed-off-by: Yu Qingyun <pinyue007@126.com>