sequencer: remove unnecessary variable setting
revs.pretty_given is only ever read in builtin/log.c, and nothing from
builtin/log.c is ever called from sequencer.c.  So setting this variable
cannot do anything.

This was introduced in commit 62db524779 ("rebase -i: generate the
script via rebase--helper", 2017-07-14), which used `git rev-list` even
though its commit message describes the logic as having been based on
`git log`.  Because of this, I am guessing this line was copied or
ported from part of builtin/log.c without recognizing that this line was
not doing anything and could be removed.

It's certainly not doing anything now, though, so remove it.

Signed-off-by: Elijah Newren <newren@gmail.com>