fast-import: align some error messages with `git branch`
We do not write "GIT" in all caps anymore, and we do translate error
messages.

The error messages when a ref needs to be created and it fails either
because the ref name is invalid or because the ref already exists should
be identical to the ones `git branch` emits.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>