mingw: support for trailing spaces on Windows
Since d2c84da (refuse to access paths with trailing spaces or periods),
git for Windows no longer handles paths containing trailing spaces and
periods. However, files with trailing spaces or periods can still be
handled correctly by the Win32 API.

For example, pass the file relative path, such as `touch("abc ");`, the
Win32 API will trim the trailing spaces and periods. But pass the file
absolute path, such as `touch("\\\\?\\C:\\abc ");`, it is possible to
treat the file correctly.

This patch allows checkout of files containing trailing spaces or
periods only if `core.protectNTFS` is disabled (`false`).
However, this patch does not change the handling of directories. This
is because the two paths will be merged or access will be refused even
if the above method is used.

Signed-off-by: Natsune Mochizuki <me@natsuneko.cat>