fsmonitor: Handle differences between Windows named pipe functions
CreateNamedPipeW is perfectly happy accepting pipe names with potential,
embedded escape sequences (e.g. \b), however WaitNamedPipeW is not and
incorrectly returns ERROR_FILE_NOT_FOUND when clearly a named pipe with
the given name exists.

For example, this network repo path is problemmatic:
\\batfs-sb29-cifs\vmgr\sbs29\mysandbox

In order to work around this issue, rather than using the path to the
worktree directly as the name of the pipe, instead use the hash of the
worktree path as the name of the pipe.

Signed-off-by: Eric DeCosta <edecosta@mathworks.com>