WIP: --find-objects: find root trees, too
TODO: add test, document that `-c` and `--cc` are buggy and show commit
headers even though the pickaxe machinery does not select the commit.

Due to technical reasons, the diff machinery does not print out the root
tree of any given commit when running with the `-t` option (for details,
see https://git-scm.com/docs/git-log#Documentation/git-log.txt--t).

However, quite often that is precisely what I need to find: the commit
that introduced a certain change that ended up with the exact
problematic (recursive) tree shape.

The underlying problem is that the layer at which `-t` prints entries is
when iterating the tree items, but the root tree is not a child of
another tree, therefore it is never reached through such an iteration.

So let's fudge things by specifically adding the root tree OID into the
`objfind` set for commits that change the top-level tree.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>